用LLVM带的clang-format plugin for Visual Studio在VS2013里尝试了一把编译, 虽然最终没有成功, 只编译成功了static lib项目, 不过也因此而发现和修正了很多小错误.

+++ src/jimic/string/csharp_sprintf.inl.h
@@ -81,7 +81,7 @@ jmc_csharp_vslprintf(jm_char *buf, size_t count, JM_CONST jm_char *fmt,
     jm_char    *end, *first;
     jm_char    *cur;
     jm_char     c;
-    size_t      index;
+    int         index;
     int         len;
 
     jimic_assert(buf != NULL);
